# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A users service API in Go using Gin framework",
        "title": "Users service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/users": {
            "post": {
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CrUserReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "422": {
                        "description": "Unprocessable",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "description": "Get user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    }
                }
            },
            "put": {
                "description": "Change user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Change user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Change successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    },
                    "422": {
                        "description": "Unprocessable",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ChUserReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.CrUserReq": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.UserResp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}